Fixes #4 - Packed Princess scenes were missing.
The difference that I had on side 2 of the disk starting at
track 26 turned out to be the packed princess scenes for that side
of the disk.

At this point, I am just left with track 17 on side 2 that I am
unsure of.

Thanks to BLuRry for originally reporting this problem.