Initial notes from RW18 investigation.
I have written up an overview of what I think the RW18 disk format
looks like on disk.  This should help me implement RW18 support in
crackle.

I am also adding the sources for RW18 routines which should work
for 5.25" disks.  It is based on the disassembly I conducted on the
RW18 object code from another RW18 based game.  I will need to move
the ORG address to $d000 for PoP when I am ready to try integrating it
in for a 5.25" disk iamge.