Make Prince of Persia run from 5.25" disk images.
Updated the crackle disk imaging tool and build process to support the
production of 5.25" disk images in addition to the 3.5" support we
already had.

Updated Makefile to create another 2 disk images.  One for each side of
the 5.25" disk image.

Added Other/PrinceOfPersia_5.25*.layout files to tell crackle how to
put together the files to create bootable 5.25" disk images.  Only the
boot loader and RW18 routines need to be placed on the disk
differently.  All of the other object and media files are placed
exactly the same as they were for the existing 3.5" layout.

Update RW18525.S to move the code up to $D000 where it is required by
Prince of Persia.

I also updated some of my notes to indicate that 5.25" disk images are
now supported.

NOTE: I have made no attempt to have these 5.25" disk images match the
      originally shipped disks.  I just concentrated on getting
      playable images.  The main reason being that I don't have access
      to an original disk to determine what the differences are and
      how they should be corrected.